+++ src/app/shared/PersonNotes.svelte
@@ -1,7 +1,6 @@
 <script lang="ts">
   import {noteKinds} from "src/util/nostr"
   import Feed from "src/app/shared/Feed.svelte"
-  import {sessions} from "src/engine"
 
   export let pubkey
   export let relays
@@ -9,4 +8,4 @@
   const filter = {kinds: noteKinds, authors: [pubkey]}
 </script>
 
-<Feed invertColors queryCache={Boolean($sessions[pubkey])} {relays} {filter} />
+<Feed invertColors {relays} {filter} />