+++ src/app/shared/RelayFeed.svelte
@@ -6,7 +6,6 @@
   import Note from "src/app/shared/Note.svelte"
 
   export let depth
-  export let showContext = false
   export let feedRelay
   export let notes
 
@@ -29,7 +28,7 @@
     <!-- If someone clicks on a child note that was seen on a relay the parent was not
          seen on, we get nothing, so just show everything - but pass down the filter -->
     {#each filteredNotes.length > 0 ? filteredNotes : notes as note (note.id)}
-      <Note invertColors {depth} {note} {feedRelay} {showContext} />
+      <Note invertColors {depth} {note} {feedRelay} />
     {/each}
   </div>