+++ src/app/views/NotificationMention.svelte
@@ -1,4 +1,5 @@
 <script lang="ts">
+  import {pluck} from "ramda"
   import {formatTimestamp} from "src/util/misc"
   import {findReplyId} from "src/util/nostr"
   import Note from "src/app/shared/Note.svelte"
@@ -9,13 +10,7 @@
 
   const {timestamp, interactions} = notification
   const parentId = findReplyId(interactions[0])
-
-  let note
-  if (parentId) {
-    note = {id: parentId, replies: interactions}
-  } else {
-    note = interactions[0]
-  }
+  const note = parentId ? {id: parentId} : interactions[0]
 </script>
 
 <div class="flex flex-col gap-4">
@@ -23,5 +18,11 @@
     <NotificationPeople {notification} actionText="mentioned you" />
     <small>{formatTimestamp(timestamp)}</small>
   </div>
-  <Note showLoading topLevel depth={1} {note} />
+  <Note
+    showLoading
+    topLevel
+    depth={1}
+    {note}
+    context={interactions}
+    filters={[{ids: pluck("id", interactions)}]} />
 </div>