+++ src/app/views/NotificationReplies.svelte
@@ -1,15 +1,13 @@
 <script lang="ts">
+  import {pluck} from "ramda"
   import {formatTimestamp} from "src/util/misc"
-  import {noteKinds, LOCAL_RELAY_URL} from "src/util/nostr"
   import NotificationPeople from "src/app/shared/NotificationPeople.svelte"
   import Note from "src/app/shared/Note.svelte"
   import type {Notification} from "src/engine"
-  import {sortEventsDesc} from "src/engine"
 
   export let notification: Notification
 
-  const {event: note, interactions, timestamp} = notification
-  const replies = sortEventsDesc(interactions.filter(e => noteKinds.includes(e.kind)))
+  const {event, interactions, timestamp} = notification
 </script>
 
 <div class="flex flex-col gap-4">
@@ -17,5 +15,10 @@
     <NotificationPeople {notification} actionText="replied to your note" />
     <small>{formatTimestamp(timestamp)}</small>
   </div>
-  <Note topLevel depth={1} relays={[LOCAL_RELAY_URL]} note={{...note, replies}} />
+  <Note
+    topLevel
+    depth={1}
+    note={event}
+    context={interactions}
+    filters={[{ids: pluck("id", interactions)}]} />
 </div>