+++ src/engine/network/utils/targets/local.ts
@@ -2,7 +2,7 @@ import {sleep} from "hurdak"
 import {Emitter} from "paravel"
 import {LOCAL_RELAY_URL} from "src/util/nostr"
 import {events} from "src/engine/events/derived"
-import {matchFilters} from "../filters"
+import {matchFilters, hasValidSignature} from "../filters"
 
 export class LocalTarget extends Emitter {
   constructor() {
@@ -16,22 +16,27 @@ export class LocalTarget extends Emitter {
   }
 
   async onREQ(subId, ...filters) {
+    const {cache, getKey} = hasValidSignature
+
     // Make sure this is async, listeners don't otherwise get attached
     await sleep(10)
 
+    const tryEvent = event => {
+      if (event && matchFilters(filters, event)) {
+        // Make sure we're not wasting time validating signatures
+        cache.set(getKey([event]), true)
+
+        this.emit("EVENT", LOCAL_RELAY_URL, subId, event)
+      }
+    }
+
     if (filters.length === 1 && filters[0].ids) {
       for (const id of filters[0].ids) {
-        const event = events.key(id).get()
-
-        if (event) {
-          this.emit("EVENT", LOCAL_RELAY_URL, subId, event)
-        }
+        tryEvent(events.key(id).get())
       }
     } else {
       for (const event of events.get()) {
-        if (matchFilters(filters, event)) {
-          this.emit("EVENT", LOCAL_RELAY_URL, subId, event)
-        }
+        tryEvent(event)
       }
     }
   }